% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression.R
\name{Lmr}
\alias{Lmr}
\title{linear_regression}
\usage{
Lmr(X, y)
}
\arguments{
\item{X}{input predictor matrix(not design matrix)}

\item{y}{continuous outcome of interest}
}
\value{
the estimated coefficient(beta)
}
\description{
Find the estimated coefficients of the linear regression model.
}
\examples{
X <- cbind(rnorm(10,1,3),rnorm(10,1,2),rnorm(10,1,1))
y <- rnorm(10,0,1)
Lmr(X,y)

}
